Include libopus as submodule
Git is not perfect regarding submodules, in fact its ui for them sucks
badly, but they seem like the best thing for our use case:
We don't want the entire history of opus in our crate (so no subtree),
nor do we want to stop updating from upstream as long as we still
perform the rewrite (so not just a simple copy of all files).

The smaller files will be rewritten instantaneously, the bigger ones
will be copied outside of the submodule into a "workbench" directory
where they will be ported function by function, getting smaller in the
process.