Add makefile that runs testsuite faster than the script
The makefile can run the testsuite in parallel, and is
therefore faster.

The improvement is pretty great: The script takes
about 2 minutes 20 seconds, while executing the
makefile with 8 jobs takes 20-40 seconds on my machine.

However, printing everything would be a mess as it would
print stuff in parallel. So we store everything in log
files. With log files however, finding out the error
message is inconvenient as it doesn't print it.
Therefore we don't remove the shell script for now, or
switch it to use the makefile.